fixed memory leak.
The issue was in the final piece of displaying the frame. I had the post-openCV image being displayed in a UIView.  I was ADDING it as subview from every frame....adding does not remove previous subviews so memory ran out.

Instead, changed preview UIView to UIImageView and now able to set its .image prproperty to each new image from OpenCV.